# Row `AVG` accumulator support Decimal type