--- a/datafusion/row/src/layout.rs
+++ b/datafusion/row/src/layout.rs
@@ -164,11 +164,15 @@ fn word_aligned_offsets(null_width: usize, schema: &Schema) -> (Vec<usize>, usiz
     let mut offset = null_width;
     for f in schema.fields() {
         offsets.push(offset);
-        assert!(!matches!(f.data_type(), DataType::Decimal128(_, _)));
-        // All of the current support types can fit into one single 8-bytes word.
-        // When we decide to support Decimal type in the future, its width would be
-        // of two 8-bytes words and should adapt the width calculation below.
-        offset += 8;
+        assert!(!matches!(f.data_type(), DataType::Decimal256(_, _)));
+        // All of the current support types can fit into one single 8-bytes word except for Decimal128.
+        // For Decimal128, its width is of two 8-bytes words.
+        match f.data_type() {
+            DataType::Decimal128(_, _) => {
+                offset += 16
+            }
+            _ => offset += 8
+        }
     }
     (offsets, offset - null_width)
 }
@@ -241,6 +245,7 @@ fn supported_type(dt: &DataType, row_type: RowType) -> bool {
                     | Float64
                     | Date32
                     | Date64
+                    | Decimal128(_, _)
             )
         }
     }